more refactors + `pkgroll` added
