+++ package.json
@@ -12,27 +12,40 @@
     "LICENSE",
     "README.md"
   ],
-  "bin": "dist/bin/index.js",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "homepage": "https://github.com/erhant/circomkit#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/erhant/circomkit.git"
-  },
-  "bugs": {
-    "url": "https://github.com/erhant/circomkit/issues/"
+  "bin": "./dist/bin/index.js",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.cts",
+  "exports": {
+    "require": {
+      "types": "./dist/index.d.cts",
+      "default": "./dist/index.cjs"
+    },
+    "import": {
+      "types": "./dist/index.d.mts",
+      "default": "./dist/index.mjs"
+    }
   },
   "scripts": {
-    "build": "npx tsc -p tsconfig.build.json",
+    "build": "pkgroll",
     "prebuild": "rimraf dist/",
     "cli": "yarn build && node ./dist/bin",
     "test": "npx mocha",
-    "pretest": "rimraf build/",
     "check": "npx tsc --noEmit",
     "style": "yarn format && yarn lint",
     "lint": "npx eslint '**/*.ts' && echo 'All good!'",
     "format": "npx prettier --check ./src/**/*.ts ./tests/**/*.ts"
+  },
+  "dependencies": {
+    "chai": "^4.3.7",
+    "@types/chai": "^4.3.4",
+    "circom_tester": "^0.0.19",
+    "loglevel": "^1.8.1",
+    "pkgroll": "^2.0.2"
+  },
+  "peerDependencies": {
+    "snarkjs": "^0.7.x",
+    "@types/snarkjs": "^0.7.8"
   },
   "devDependencies": {
     "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
@@ -44,11 +57,9 @@
     "@nomicfoundation/hardhat-verify": "^2.0.0",
     "@typechain/ethers-v6": "^0.5.0",
     "@typechain/hardhat": "^9.0.0",
-    "@types/chai": "^4.3.4",
     "@types/mocha": "^10.0.1",
     "@types/mocha-each": "^2.0.0",
     "@types/node": "^18.11.18",
-    "@types/snarkjs": "^0.7.8",
     "@typescript-eslint/eslint-plugin": "^6.13.1",
     "@typescript-eslint/parser": "^6.13.1",
     "eslint": "^8.54.0",
@@ -68,12 +79,6 @@
     "typechain": "^8.3.0",
     "typescript": "^5.3.2"
   },
-  "dependencies": {
-    "chai": "^4.3.7",
-    "circom_tester": "^0.0.19",
-    "loglevel": "^1.8.1",
-    "snarkjs": "^0.7.0"
-  },
   "prettier": {
     "bracketSpacing": false,
     "singleQuote": true,
@@ -89,5 +94,13 @@
     "cli",
     "tooling",
     "blockchain"
-  ]
+  ],
+  "homepage": "https://github.com/erhant/circomkit#readme",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/erhant/circomkit.git"
+  },
+  "bugs": {
+    "url": "https://github.com/erhant/circomkit/issues/"
+  }
 }
