+++ src/index.ts
@@ -1,6 +1,6 @@
 export {Circomkit} from './circomkit';
 
-// TODO: make these type exports only on next minor version
+// @todo make these type exports only on next minor version
 export {ProofTester, WitnessTester} from './testers/';
 export type {CircuitSignals, CircuitConfig} from './types/circuit';
 export type {CircomkitConfig} from './types/circomkit';
