+++ tests/inputs/multiplier_3/test-input.json
@@ -1,7 +1,7 @@
 {
   "in": [
-    "10190322864210778638",
-    "6513347404669051035",
-    "9099428092265999537"
+    "6610721603614245176",
+    "16320090575426234174",
+    "14295824264710608063"
   ]
 }
