--- tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "noEmit": false,
-    "skipLibCheck": true
-  },
-  "include": ["src/**/*.ts"]
-}
